게시글 관련 이슈 처리
게시글 관련 이슈 처리 했습니다

+++ free_board_read_del.php
@@ -1,8 +1,17 @@
 <?php
 SESSION_START();
-$no=$_GET['no'];
+if(!isset($_SESSION['id']))
+{
+	echo "<script>
+	alert('비정상적인 접근입니다.');
+	location.href='free_board.php?page=1';
+	</script>";	
+	exit;
+}
+$no=$_POST['no'];
+$page=$_POST['page'];
 include 'conn.php';
-$board_name=$_GET['board_name'];
+$board_name=$_POST['board_name'];
 
 $sql="delete from {$board_name} where no={$no}";
 $result=mysqli_query($conn,$sql);
@@ -12,7 +21,7 @@ if($result)
 	$result=mysqli_query($conn,$sql);
 		echo "<script>
 		alert('게시글이 삭제 되었습니다.');
-		location.href='/free_board.php';
+		location.href='/free_board.php?page={$page}';
 		</script>";
 }
 else
@@ -22,5 +31,6 @@ else
 		location.href='javascript:history.back()';
 		</script>";
 }
+
 mysqli_close($conn);
 ?>
